refactor(component): migrate Footer to CSS Modules
